refactor: prepare ContentIcalElement
