+++ contao/languages/de/default.php
@@ -10,8 +10,11 @@ declare(strict_types=1);
  * @license    LGPL-3.0-or-later
  */
 
-$GLOBALS['TL_LANG']['CTE']['ical']['0'] = 'iCal';
-$GLOBALS['TL_LANG']['CTE']['ical']['1'] = 'Dieses Element erstellt einen Hyperlink auf eine vCalendar (iCal)-Datei, die heruntergeladen werden kann und dazu verwendet werden kann, die enthaltenen Termine in einer externen Kalenderanwendung anzuzeigen.';
+use Cgoit\ContaoCalendarIcalBundle\Controller\ContentElement\ContentIcalElement;
+
+$GLOBALS['TL_LANG']['CTE'][ContentIcalElement::TYPE]['0'] = 'iCal';
+$GLOBALS['TL_LANG']['CTE'][ContentIcalElement::TYPE]['1'] = 'Dieses Element erstellt einen Hyperlink auf eine vCalendar (iCal)-Datei, die heruntergeladen werden kann und dazu verwendet werden kann, die enthaltenen Termine in einer externen Kalenderanwendung anzuzeigen.';
+
 $GLOBALS['TL_LANG']['MSC']['import_calendar']['0'] = 'Kalender importieren';
 $GLOBALS['TL_LANG']['MSC']['import_calendar']['1'] = 'Importieren Sie eine iCal (.ics) Kalender-Datei in Ihren Kalender.';
 $GLOBALS['TL_LANG']['MSC']['export_calendar']['0'] = 'Kalender exportieren';
