+++ contao/languages/de/tl_content.php
@@ -10,13 +10,18 @@ declare(strict_types=1);
  * @license    LGPL-3.0-or-later
  */
 
+$GLOBALS['TL_LANG']['tl_content']['ical_download_title']['0'] = 'iCal-Datei herunterladen';
+
 $GLOBALS['TL_LANG']['tl_content']['ical_calendar']['0'] = 'Kalender';
 $GLOBALS['TL_LANG']['tl_content']['ical_calendar']['1'] = 'Bitte wählen Sie einen oder mehrere Kalender aus.';
 $GLOBALS['TL_LANG']['tl_content']['ical_start']['0'] = 'Startdatum';
 $GLOBALS['TL_LANG']['tl_content']['ical_start']['1'] = 'Bitte geben Sie das Startdatum ein, ab dem Contao Termine aus den ausgewählten Kalendern extrahieren soll. Wenn Sie kein Datum angeben, wird das aktuelle Datum als Startdatum verwendet.';
 $GLOBALS['TL_LANG']['tl_content']['ical_end']['0'] = 'Enddatum';
 $GLOBALS['TL_LANG']['tl_content']['ical_end']['1'] = 'Bitte geben Sie das Endatum ein, bis zu dem Contao Termine aus den ausgewählten Kalendern extrahieren soll.  Wenn Sie kein Datum angeben, wird das Datum in einem Jahr als Enddatum verwendet.';
-$GLOBALS['TL_LANG']['tl_content']['ical_title'] = 'iCal-Datei herunterladen';
+$GLOBALS['TL_LANG']['tl_content']['ical_title']['0'] = 'iCal Titel';
+$GLOBALS['TL_LANG']['tl_content']['ical_title']['1'] = 'Bitte geben Sie einen Titel für das iCal-Abonnement ein';
+$GLOBALS['TL_LANG']['tl_content']['ical_description']['0'] = 'iCal Beschreibung';
+$GLOBALS['TL_LANG']['tl_content']['ical_description']['1'] = 'Bitte geben Sie eine kurze Beschreibung des iCal-Abonnements ein';
 $GLOBALS['TL_LANG']['tl_content']['ical_prefix']['0'] = 'Titel-Präfix';
 $GLOBALS['TL_LANG']['tl_content']['ical_prefix']['1'] = 'Bitte geben Sie ein Präfix ein, das in der iCal-Abonnement-Datei vor allen Event-Titeln ausgegeben wird.';
 $GLOBALS['TL_LANG']['tl_content']['calendar_legend'] = 'Kalender-Einstellungen';
