+++ contao/languages/en/default.php
@@ -10,7 +10,10 @@ declare(strict_types=1);
  * @license    LGPL-3.0-or-later
  */
 
-$GLOBALS['TL_LANG']['CTE']['ical'] = ['iCal', 'This element contains a hyperlink to an vCalendar (iCal) which can be downloaded and used to display calendar dates in an external calendar application.'];
+use Cgoit\ContaoCalendarIcalBundle\Controller\ContentElement\ContentIcalElement;
+
+$GLOBALS['TL_LANG']['CTE'][ContentIcalElement::TYPE] = ['iCal', 'This element contains a hyperlink to an vCalendar (iCal) which can be downloaded and used to display calendar dates in an external calendar application.'];
+
 $GLOBALS['TL_LANG']['MSC']['import_calendar'] = ['Import calendar', 'Import an iCal (.ics) file with events into your calendar.'];
 $GLOBALS['TL_LANG']['MSC']['export_calendar'] = ['Export calendar', 'Export an iCal (.ics) file with events from your calendar.'];
 $GLOBALS['TL_LANG']['MSC']['location'] = 'Location';
