+++ contao/languages/en/tl_content.php
@@ -10,9 +10,12 @@ declare(strict_types=1);
  * @license    LGPL-3.0-or-later
  */
 
+$GLOBALS['TL_LANG']['tl_content']['ical_download_title']['0'] = 'Download iCal';
+
 $GLOBALS['TL_LANG']['tl_content']['ical_calendar'] = ['Calendars', 'Please choose one or more calendars.'];
 $GLOBALS['TL_LANG']['tl_content']['ical_start'] = ['Start date', 'Please enter the start date of the calendar. If you do not enter a date, the actual date will be taken as start date.'];
 $GLOBALS['TL_LANG']['tl_content']['ical_end'] = ['End date', 'Please enter the end date of the calendar. If you do not enter a date, the date in one year will be taken as end date.'];
-$GLOBALS['TL_LANG']['tl_content']['ical_title'] = 'Download iCal';
+$GLOBALS['TL_LANG']['tl_content']['ical_title'] = ['iCal title', 'Please enter the title of the calendar'];
+$GLOBALS['TL_LANG']['tl_content']['ical_description'] = ['iCal description', 'Please enter a short description of the calendar.'];
 $GLOBALS['TL_LANG']['tl_content']['ical_prefix'] = ['Title prefix', 'Here you can enter a prefix that will be added to every event title in the iCal subscription.'];
 $GLOBALS['TL_LANG']['tl_content']['calendar_legend'] = 'Calendar settings';
